requirements.txt
shoeboxpy
numpy
scipy
matplotlib
pandas
casadi
ipython
ipywidgets
ipympl
gymnasium
shapepy
networkx
pyyaml